requirements.txt
chatterbot==0.8.7
django==2.1.3
djangorestframework-jwt==1.11.0
djangorestframework==3.9.0
psycopg2-binary==2.7.6